libs/time: move types/time into libs
