+++ types/genesis_test.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/tendermint/tendermint/crypto/ed25519"
 	tmjson "github.com/tendermint/tendermint/libs/json"
-	tmtime "github.com/tendermint/tendermint/types/time"
+	tmtime "github.com/tendermint/tendermint/libs/time"
 )
 
 func TestGenesisBad(t *testing.T) {
